LayoutTests/fast/forms/menulist-no-renderer-for-unexpected-children.html

<!DOCTYPE html>
<script src="../../resources/js-test-pre.js"></script>
<select size="2">
  <option>PASS</option>FAIL
</select>
<script>
description('Check that a select control does not render children that are not &lt;option&gt; or &lt;optgroup&gt;.');

var select = document.querySelector('select');

var div = select.appendChild(document.createElement('div'));
div.innerText = 'FAIL';
div.style.background = 'red';

shouldBeEqualToString('select.innerText', 'PASS');
</script>
<script src="../../resources/js-test-post.js"></script>
